Chore: 컴포넌트 폴더 구조 변경 및 스타일 적용
