+++ client/src/components/Dropdown/style.scss
@@ -1,0 +1,51 @@
+.dropdown {
+	font-size: 1.6rem;
+	font-weight: 400;
+	color: #000;
+	width: 14.8rem;
+	border-radius: 1rem;
+	border: 1px solid #ccd3d3;
+	background-color: #fff;
+}
+
+.dropdown-list > li {
+	cursor: pointer;
+}
+
+.dropdown-list > li:hover {
+	background-color: #f5f5f5;
+}
+
+.dropdown-list > li:first-child {
+	border-top-left-radius: 1rem;
+	border-top-right-radius: 1rem;
+}
+
+.dropdown-list > li:last-child {
+	border: none;
+	border-bottom-left-radius: 1rem;
+	border-bottom-right-radius: 1rem;
+}
+
+.dropdown-list-item {
+	padding: 0 2.4rem;
+	text-align: left;
+	div {
+		padding: 1.6rem 0;
+		width: 10rem;
+		border-bottom: 1px solid #ccd3d3;
+	}
+}
+
+.dropdown-add {
+	padding: 1.6rem 2.4rem;
+	width: 10rem;
+	border-bottom: 1px solid #ccd3d3;
+	text-align: left;
+	display: flex;
+	justify-content: space-between;
+	img {
+		width: 1.6rem;
+		height: 1.6rem;
+	}
+}
