+++ client/src/components/InputBar/style.scss
@@ -1,0 +1,20 @@
+.inputbar {
+	position: absolute;
+	left: 50%;
+	top: 20rem;
+	width: 85.6rem;
+	height: 7.6rem;
+	transform: translate(-50%, -50%);
+	background-color: #fff;
+	text-align: center;
+	display: flex;
+	border-radius: 1rem;
+	border: 1px solid #c1c5c5;
+	padding: 0 1.6rem 0 3.2rem;
+	align-items: center;
+	box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.1), 0px 4px 20px rgba(0, 0, 0, 0.1);
+}
+
+.inputbar-save {
+	cursor: pointer;
+}
