+++ client/src/components/InputItem/style.scss
@@ -1,0 +1,85 @@
+.inputbar-item {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: flex-start;
+	border-right: 1px solid #c1c5c5;
+	padding: 0.5rem 2.4rem;
+}
+
+.inputbar-item-first {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: flex-start;
+	border-right: 1px solid #c1c5c5;
+	padding: 0.5rem 0;
+	padding-right: 2.4rem;
+}
+
+.inputbar-item-last {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: flex-start;
+	border: none;
+	padding-left: 2.4rem;
+	width: 20.8rem;
+}
+
+.inputbar-label {
+	font-family: 'Noto Sans KR';
+	font-size: 1.2rem;
+	font-weight: 700;
+	color: #219a95;
+	padding-bottom: 0.4rem;
+}
+
+.inputbar-input {
+	font-size: 1.6rem;
+}
+
+.inputbar-date {
+	font-size: 1.6rem;
+	border: none;
+	width: 8rem;
+}
+
+.inputbar-category > select {
+	font-size: 1.6rem;
+	border: none;
+	margin-left: -0.5rem;
+	padding-right: 0.8rem;
+}
+
+.inputbar-dropdown:first-child {
+	color: #757575;
+}
+
+.inputbar-content {
+	font-size: 1.6rem;
+	border: none;
+	width: 17.2rem;
+}
+
+.inputbar-payment > select {
+	font-size: 1.6rem;
+	border: none;
+	margin-left: -0.5rem;
+	padding-right: 0.8rem;
+}
+
+.inputbar-io {
+	position: relative;
+	top: 0.3rem;
+	padding-right: 1.1rem;
+	cursor: pointer;
+}
+
+.inputbar-amount {
+	width: 9.6rem;
+	font-size: 1.6rem;
+	border: none;
+	text-align: right;
+	padding-right: 0.5rem;
+}
