+++ client/src/components/MainHeader/style.scss
@@ -1,0 +1,58 @@
+.header {
+	font-family: 'BMDOHYEON';
+	font-weight: 400;
+	background-color: #2ac1bc;
+	height: 19.4rem;
+	color: white;
+	display: flex;
+	align-items: center;
+	justify-content: space-around;
+	padding: 0 24rem;
+}
+
+.header-title {
+	font-size: 2.4rem;
+	line-height: 2.4rem;
+	cursor: pointer;
+}
+
+.header-row {
+	display: flex;
+	justify-items: center;
+	align-items: center;
+	span {
+		img {
+			width: 2.4rem;
+			height: 2.4rem;
+			padding: 2.6rem;
+			cursor: pointer;
+		}
+	}
+}
+
+.header-date {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	padding: 0 3rem;
+}
+
+.header-month {
+	font-size: 4.8rem;
+	padding-bottom: 1rem;
+}
+
+.header-year {
+	font-size: 2.4rem;
+}
+
+.header-tab {
+	span {
+		img {
+			width: 2.4rem;
+			height: 2.4rem;
+			margin: 0 1rem;
+			cursor: pointer;
+		}
+	}
+}
