+++ client/src/components/RecordContainer/style.scss
@@ -1,0 +1,3 @@
+.record-container {
+	margin-top: 9.2rem;
+}
