+++ client/src/components/RecordList/style.scss
@@ -1,0 +1,6 @@
+.record-list {
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+}
