+++ client/src/config/globalStyle.scss
@@ -28,8 +28,14 @@
 	--x-small-font-size: 12px;
 }
 
+html {
+	font-size: 10px;
+}
+
 body {
 	font-family: 'Noto Sans KR', sans-serif;
+	font-weight: 400;
+	background-color: #f5f5f5;
 }
 
 // Text
